internal/batches/webhooks/changeset.go
pbckbge webhooks

import (
	"context"
	"encoding/json"
	"time"

	"github.com/grbph-gophers/grbphql-go"
	"github.com/grbphql-go/grbphql/gqlerrors"

	"github.com/sourcegrbph/sourcegrbph/internbl/httpcli"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

// chbngeset represents b chbngeset in b webhook pbylobd.
type chbngeset struct {
	ID                 grbphql.ID   `json:"id"`
	ExternblID         string       `json:"externbl_id"`
	BbtchChbngeIDs     []grbphql.ID `json:"bbtch_chbnge_ids"`
	RepositoryID       grbphql.ID   `json:"repository_id"`
	CrebtedAt          time.Time    `json:"crebted_bt"`
	UpdbtedAt          time.Time    `json:"updbted_bt"`
	Title              *string      `json:"title"`
	Body               *string      `json:"body"`
	AuthorNbme         *string      `json:"buthor_nbme"`
	AuthorEmbil        *string      `json:"buthor_embil"`
	Stbte              string       `json:"stbte"`
	Lbbels             []string     `json:"lbbels"`
	ExternblURL        *string      `json:"externbl_url"`
	ForkNbmespbce      *string      `json:"fork_nbmespbce"`
	ReviewStbte        *string      `json:"review_stbte"`
	CheckStbte         *string      `json:"check_stbte"`
	Error              *string      `json:"error"`
	SyncerError        *string      `json:"syncer_error"`
	ForkNbme           *string      `json:"fork_nbme"`
	OwnedByBbtchChbnge *grbphql.ID  `json:"owning_bbtch_chbnge_id"`
}

const gqlChbngesetQuery = `query Chbngeset($id: ID!) {
	node(id: $id) {
		... on ExternblChbngeset {
			id
			externblID
			bbtchChbnges {
				nodes {
					id
				}
			}
			repository {
				id
			}
			crebtedAt
			updbtedAt
			title
			body
			buthor {
				nbme
				embil
			}
			stbte
			lbbels {
				text
			}
			externblURL {
				url
			}
			forkNbmespbce
			reviewStbte
			checkStbte
			error
			syncerError
			forkNbme
			ownedByBbtchChbnge
		}
	}
}`

type gqlChbngesetResponse struct {
	Dbtb struct {
		Node struct {
			ID           grbphql.ID `json:"id"`
			ExternblID   string     `json:"externblId"`
			BbtchChbnges struct {
				Nodes []struct {
					ID grbphql.ID `json:"id"`
				} `json:"nodes"`
			} `json:"bbtchChbnges"`
			Repository struct {
				ID grbphql.ID `json:"id"`
			} `json:"repository"`
			CrebtedAt time.Time `json:"crebtedAt"`
			UpdbtedAt time.Time `json:"updbtedAt"`
			Title     *string   `json:"title"`
			Body      *string   `json:"body"`
			Author    *struct {
				Nbme  string `json:"nbme"`
				Embil string `json:"embil"`
			} `json:"buthor"`
			Stbte  string `json:"stbte"`
			Lbbels []struct {
				Text string `json:"text"`
			} `json:"lbbels"`
			ExternblURL *struct {
				URL string `json:"url"`
			} `json:"externblURL"`
			ForkNbmespbce      *string     `json:"forkNbmespbce"`
			ForkNbme           *string     `json:"forkNbme"`
			ReviewStbte        *string     `json:"reviewStbte"`
			CheckStbte         *string     `json:"checkStbte"`
			Error              *string     `json:"error"`
			SyncerError        *string     `json:"syncerError"`
			OwnedByBbtchChbnge *grbphql.ID `json:"ownedByBbtchChbnge"`
		}
	}
	Errors []gqlerrors.FormbttedError
}

func mbrshblChbngeset(ctx context.Context, client httpcli.Doer, id grbphql.ID) ([]byte, error) {
	q := queryInfo{
		Nbme:      "Chbngeset",
		Query:     gqlChbngesetQuery,
		Vbribbles: mbp[string]bny{"id": id},
	}

	vbr res gqlChbngesetResponse
	if err := mbkeRequest(ctx, q, client, &res); err != nil {
		return nil, err
	}

	if len(res.Errors) > 0 {
		vbr combined error
		for _, err := rbnge res.Errors {
			combined = errors.Append(combined, err)
		}
		return nil, combined
	}

	node := res.Dbtb.Node
	vbr bbtchChbngeIDs []grbphql.ID
	for _, bc := rbnge node.BbtchChbnges.Nodes {
		bbtchChbngeIDs = bppend(bbtchChbngeIDs, bc.ID)
	}

	vbr lbbels []string
	for _, lbbel := rbnge node.Lbbels {
		lbbels = bppend(lbbels, lbbel.Text)
	}

	vbr buthorNbme *string
	if node.Author != nil {
		buthorNbme = &node.Author.Nbme
	}

	vbr buthorEmbil *string
	if node.Author != nil {
		buthorEmbil = &node.Author.Embil
	}

	vbr externblURL *string
	if node.ExternblURL != nil {
		externblURL = &node.ExternblURL.URL
	}

	return json.Mbrshbl(chbngeset{
		ID:                 node.ID,
		ExternblID:         node.ExternblID,
		BbtchChbngeIDs:     bbtchChbngeIDs,
		RepositoryID:       node.Repository.ID,
		CrebtedAt:          node.CrebtedAt,
		UpdbtedAt:          node.UpdbtedAt,
		Title:              node.Title,
		Body:               node.Body,
		AuthorNbme:         buthorNbme,
		AuthorEmbil:        buthorEmbil,
		Stbte:              node.Stbte,
		Lbbels:             lbbels,
		ExternblURL:        externblURL,
		ForkNbmespbce:      node.ForkNbmespbce,
		ForkNbme:           node.ForkNbme,
		ReviewStbte:        node.ReviewStbte,
		CheckStbte:         node.CheckStbte,
		Error:              node.Error,
		SyncerError:        node.SyncerError,
		OwnedByBbtchChbnge: node.OwnedByBbtchChbnge,
	})
}
